+++ semcore/input-number/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [3.0.0] - unreleased
+
+### BREAK
+
+- Replaced internal representation with native input(type=number).
+- Changed type for value to string.
+
 ## [2.2.0] - 2021-04-26
 
 ### Changed